# api/openapi/kcsp-v0.1.yaml
openapi: 3.1.0
info:
  title: KCSP Embedded Vertical Slice API
  version: 0.1.0
  summary: Executable development contract for the Event-to-Incident slice.
  description: |
    This contract describes the executable `embedded-dev` profile. The API runs one
    synchronous ingest/detection/SOC pipeline and stores state in process memory.

    Authentication and authorization use hard-coded demonstration Bearer tokens plus
    the required `X-KCSP-Tenant-ID` membership check. These tokens are public test data,
    not secrets and not production OIDC. Runtime state and the SHA-256 audit chain are
    lost on restart. The profile is not suitable for real telemetry or an untrusted
    network.

    Kafka, ClickHouse, PostgreSQL, and OIDC are intended future adapters behind ports;
    they are not implemented by V0.1.
  contact:
    name: KCSP engineering
  license:
    name: Proprietary - development contract
    identifier: LicenseRef-KCSP-Development
servers:
  - url: http://127.0.0.1:8080
    description: Loopback-published embedded development API
security:
  - demoBearer: []
...
paths:
  /api/v1/session:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Session]
      operationId: getSession
      summary: Return the effective demo principal and tenant context.
      x-kcsp-permission: platform.overview.read
      responses:
        '200':
          description: Effective local session.
          content:
            application/json:
              schema:
                type: object
                required: [principal, tenant, permissions, locale, timezone, auth_mode]
                properties:
                  principal:
                    type: object
                    required: [id, display_name, role]
                    properties:
                      id: { type: string }
                      display_name: { type: string }
                      role: { type: string }
                  tenant:
                    type: object
                    required: [id, name]
                    properties:
                      id: { type: string }
                      name: { type: string }
                  permissions:
                    type: array
                    items: { type: string }
                  locale: { type: string, example: ru-KZ }
                  timezone: { type: string, example: Asia/Qyzylorda }
                  auth_mode: { type: string, const: demo }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/demo/reset:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    post:
      tags: [Demo]
      operationId: resetDemoProfile
      summary: Replace volatile runtime state with the synthetic seed baseline.
      description: Development-only endpoint protected by the platform-admin fixture token.
      x-kcsp-permission: platform.demo.reset
      responses:
        '200':
          description: Seed baseline restored.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, const: reset }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/overview:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Overview]
      operationId: getOverview
      summary: Return the current tenant dashboard summary.
      x-kcsp-permission: platform.overview.read
      responses:
        '200':
          description: Overview generated from current in-memory state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Overview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Events]
      operationId: listEvents
      summary: List tenant events in reverse insertion order.
      x-kcsp-permission: siem.events.read
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: category
          in: query
          schema: { type: string }
        - name: severity
          in: query
          schema: { type: integer }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Event list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Events]
      operationId: ingestEvent
      summary: Run one canonical event through the synchronous embedded pipeline.
      description: |
        The caller submits the canonical V0.1 envelope directly; V0.1 does not parse
        raw Sysmon/XML/CEF input. The server overwrites `tenant_id`, `ingest_time`, and
        `raw.hash`, and supplies defaults for missing event id, event time, collector,
        schema version, parser, raw message, and raw reference.

        A new event returns 201 with separate Event, Finding, and Alert objects. An
        existing `event_id` in the same server-bound tenant returns 200 with
        `duplicate: true` and empty Finding/Alert arrays.
      x-kcsp-permission: siem.events.ingest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanonicalEventInput'
            examples:
              suspiciousPowerShell:
                summary: Synthetic canonical Sysmon process-creation event
                value:
                  event_id: fixture-sysmon-positive-001
                  category: process_activity
                  activity_name: Process created
                  source:
                    vendor: Microsoft
                    product: Sysmon
                    type: endpoint
                  user:
                    name: 'KUTB\administrator'
                    is_privileged: true
                  device:
                    hostname: dc-01
                    criticality: 5
                  process:
                    name: powershell.exe
                    command_line: powershell.exe -NoP -EncodedCommand SQBFAFgA
      responses:
        '201':
          description: A new Event was processed synchronously.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '200':
          description: The tenant already contains this `event_id`; no new detection objects were created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events/{eventID}:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      operationId: getEvent
      summary: Return one immutable canonical Event.
      x-kcsp-permission: siem.events.read
      responses:
        '200':
          description: Event found in the server-bound tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanonicalEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/findings:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Findings]
      operationId: listFindings
      summary: List immutable Findings, optionally for one Event.
      x-kcsp-permission: siem.findings.read
      parameters:
        - name: event_id
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Finding list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/alerts:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Alerts]
      operationId: listAlerts
      summary: List tenant Alerts requiring or recording triage.
      x-kcsp-permission: soc.alerts.read
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AlertStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/Severity'
        - name: assignee
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Alert list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/alerts/{alertID}:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
      - $ref: '#/components/parameters/AlertId'
    get:
      tags: [Alerts]
      operationId: getAlert
      summary: Return one Alert with ETag version.
      x-kcsp-permission: soc.alerts.read
      responses:
        '200':
          description: Alert found.
          headers:
            ETag:
              $ref: '#/components/headers/VersionETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Alerts]
      operationId: updateAlert
      summary: Triage an Alert.
      description: |
        Supply the expected optimistic version either in `version` or `If-Match`.
        A stale version returns 412. Closing requires a disposition. Unknown JSON
        fields are rejected.
      x-kcsp-permission: soc.alerts.manage
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertPatch'
      responses:
        '200':
          description: Alert updated and an AuditEntry appended.
          headers:
            ETag:
              $ref: '#/components/headers/VersionETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/incidents:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Incidents]
      operationId: listIncidents
      summary: List tenant Incidents.
      x-kcsp-permission: soc.incidents.read
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/IncidentStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/Severity'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Incident list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Incidents]
      operationId: createIncident
      summary: Create an Incident from one or more Alerts.
      description: |
        Severity, risk, Event/Finding lineage, entities, and MITRE techniques are
        derived from the referenced Alerts. Repeating the same normalized Alert set
        returns the existing Incident with 200.
      x-kcsp-permission: soc.incidents.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
      responses:
        '201':
          description: Incident created and AuditEntry appended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '200':
          description: An Incident already exists for exactly this Alert set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/rules:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Rules]
      operationId: listRules
      summary: List the two embedded, read-only V0.1 detection rules.
      x-kcsp-permission: detection.rules.read
      responses:
        '200':
          description: Rule list containing PowerShell and authentication-threshold rules.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/audit:
    parameters:
      - $ref: '#/components/parameters/TenantHeader'
    get:
      tags: [Audit]
      operationId: listAuditEntries
      summary: List volatile per-tenant SHA-256 chained AuditEntries.
      description: |
        `chain_valid` recomputes the in-memory chain. The chain is unkeyed, volatile,
        and not backed by WORM storage, so it is not forensic tamper evidence.
      x-kcsp-permission: platform.audit.read
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Audit list and current chain verification result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    demoBearer:
      type: http
      scheme: bearer
      bearerFormat: KCSP-DEMO-TOKEN
      description: |
        Hard-coded development identity. Examples include `kcsp-demo-collector`,
        `kcsp-demo-l1`, `kcsp-demo-l2`, `kcsp-demo-auditor`, and `kcsp-demo-admin`.
        These values are public fixtures and must never be enabled in production.
  parameters:
    TenantHeader:
      name: X-KCSP-Tenant-ID
      in: header
      required: true
      description: Tenant requested by the caller and authorized against demo-token membership.
      schema:
        type: string
        minLength: 1
      example: university-kulazhanov
    Query:
      name: q
      in: query
      description: Case-insensitive in-memory substring search over resource-specific fields.
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing or unknown demo Bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Token lacks the operation permission or requested tenant membership.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InternalError:
      description: Unexpected embedded adapter failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...